Rendering system rewrite (~ one year old changes)
Separate GMWindow and Scene, such that the scene doesn't depend on
GMWindow.
Render directly to resource (e.g.: texturebuffer or rbo).
All objects now lives in the scene, none in the window.

Stage #1

+++ src/scene/render/gmrendertarget.h
@@ -13,22 +13,22 @@ public:
   RenderTarget() : _size() {}
   virtual ~RenderTarget() {}
 
-  void                      clear() const;
-  void                      bind() const;
-  void                      unbind() const;
+  void                              clear() const;
+  void                              bind() const;
+  void                              unbind() const;
 
-  const Vector<float,2>&    getSize() const;
-  const float               getWidth() const;
-  const float               getHeight() const;
-  void                      resize( const Vector<float,2>& size );
+  const Vector<unsigned int,2>&     getSize() const;
+  const float                       getWidth() const;
+  const float                       getHeight() const;
+  void                              resize( const Vector<unsigned int,2>& size );
 
 private:
-  Vector<float,2>     _size;
+  Vector<unsigned int,2>            _size;
 
-  virtual void        doClear() const = 0;
-  virtual void        doBind() const = 0;
-  virtual void        doUnbind() const = 0;
-  virtual void        doResize() = 0;
+  virtual void                      doClear() const = 0;
+  virtual void                      doBind() const = 0;
+  virtual void                      doUnbind() const = 0;
+  virtual void                      doResize() = 0;
 }; // END class RenderTarget
 
 
@@ -45,7 +45,7 @@ inline
 void RenderTarget::unbind() const { doUnbind(); }
 
 inline
-const Vector<float,2>&RenderTarget::getSize() const { return _size; }
+const Vector<unsigned int,2>&RenderTarget::getSize() const { return _size; }
 
 inline
 const float RenderTarget::getWidth() const { return _size(0); }
@@ -54,7 +54,7 @@ inline
 const float RenderTarget::getHeight() const { return _size(1); }
 
 inline
-void RenderTarget::resize(const Vector<float,2>& size)  { _size = size; doResize(); }
+void RenderTarget::resize(const Vector<unsigned int,2>& size)  { _size = size; doResize(); }
 
 
 } // End namespace GMlib
